pyproject.toml
[project]
name = "carrotquant-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Carrot", email = "11172391+CRThu@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "akshare>=1.18.39",
    "fastapi>=0.135.1",
    "loguru>=0.7.3",
    "polars>=1.39.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0.3",
    "typer>=0.24.1",
    "uvicorn>=0.41.0",
]

[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"
